Make a new generic OnRPCListener
