+++ sdl_android/src/main/java/com/smartdevicelink/api/ProxyBridge.java
@@ -76,6 +76,7 @@ import com.smartdevicelink.proxy.rpc.UnsubscribeVehicleDataResponse;
 import com.smartdevicelink.proxy.rpc.UnsubscribeWayPointsResponse;
 import com.smartdevicelink.proxy.rpc.UpdateTurnListResponse;
 import com.smartdevicelink.proxy.rpc.enums.SdlDisconnectedReason;
+import com.smartdevicelink.proxy.rpc.listeners.OnRPCListener;
 
 import java.util.concurrent.CopyOnWriteArrayList;
 
@@ -105,10 +106,6 @@ public class ProxyBridge implements IProxyListener{
 		onRPCReceived(response);
 	}
 
-	public interface OnRPCListener {
-		void onRpcReceived(int functionID, RPCMessage message);
-	}
-
 	protected interface LifecycleListener{
 		void onProxyConnected();
 		void onProxyClosed(String info, Exception e, SdlDisconnectedReason reason);
@@ -128,7 +125,7 @@ public class ProxyBridge implements IProxyListener{
 			CopyOnWriteArrayList<OnRPCListener> listeners = rpcListeners.get(id);
 			if(listeners!=null && listeners.size()>0) {
 				for (OnRPCListener listener : listeners) {
-					listener.onRpcReceived(id,message);
+					listener.onReceived(message);
 				}
 				return true;
 			}
