+++ sdl_android/src/main/java/com/smartdevicelink/proxy/rpc/listeners/OnRPCResponseListener.java
@@ -1,46 +1,20 @@
 package com.smartdevicelink.proxy.rpc.listeners;
 
+import com.smartdevicelink.proxy.RPCMessage;
 import com.smartdevicelink.proxy.RPCResponse;
 import com.smartdevicelink.proxy.rpc.enums.Result;
 
-public abstract class OnRPCResponseListener {
-	/**
-	 * Generic listener type that will work for most RPCs
-	 */
-	public final static int UPDATE_LISTENER_TYPE_BASE_RPC 		= 0;
-	/**
-	 * Listener type specific to putfile
-	 */
-	public final static int UPDATE_LISTENER_TYPE_PUT_FILE 		= 1;
-	/**
-	 * Listener type specific to sendRequests and sendSequentialRequests
-	 */
-	public final static int UPDATE_LISTENER_TYPE_MULTIPLE_REQUESTS 		= 2;
+public abstract class OnRPCResponseListener extends OnRPCListener {
 
-	/**
-	 * Stores what type of listener this instance is. This prevents of from having to use reflection
-	 */
-	int listenerType;
-	
-	/**
-	 * This is the base listener for all RPCs.
-	 */
+
 	public OnRPCResponseListener(){
-		setListenerType(UPDATE_LISTENER_TYPE_BASE_RPC);
+		setListenerType(OnRPCListener.UPDATE_LISTENER_TYPE_BASE_RPC);
 	}
-	
-	protected final void setListenerType(int type){
-		this.listenerType = type;
+
+	public final void onReceived(final RPCMessage message){
+		//Do nothing
 	}
-	/**
-	 * This is used to see what type of listener this instance is. It is needed
-	 * because some RPCs require additional callbacks. Types are  constants located in this class
-	 * @return the type of listener this is 
-	 */
-	public int getListenerType(){
-		return this.listenerType;
-	}
-	
+
 	/* *****************************************************************
 	 ************************* Event Callbacks *************************
 	 *******************************************************************/
